Move opengever.ogds.models.types to opengever.base.
