+++ opengever/base/types.py
