+++ opengever/contact/models/contact.py
@@ -1,7 +1,7 @@
 from opengever.base.model import Base
 from opengever.base.model import SQLFormSupport
 from opengever.contact.models.participation import ContactParticipation
-from opengever.ogds.models.types import UnicodeCoercingText
+from opengever.base.types import UnicodeCoercingText
 from sqlalchemy import Boolean
 from sqlalchemy import Column
 from sqlalchemy import Integer
