+++ opengever/contact/upgrades/20160630155228_add_sql_model_for_new_contact_implementation/upgrade.py
@@ -1,5 +1,5 @@
 from opengever.core.upgrade import SchemaMigration
-from opengever.ogds.models.types import UnicodeCoercingText
+from opengever.base.types import UnicodeCoercingText
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
