+++ opengever/globalindex/model/task.py
@@ -13,7 +13,7 @@ from opengever.ogds.base.utils import ogds_service
 from opengever.ogds.models import UNIT_ID_LENGTH
 from opengever.ogds.models import USER_ID_LENGTH
 from opengever.base.query import BaseQuery
-from opengever.ogds.models.types import UnicodeCoercingText
+from opengever.base.types import UnicodeCoercingText
 from opengever.task.reminder.reminder import TaskReminder
 from plone import api
 from Products.CMFPlone.utils import safe_unicode
