+++ opengever/meeting/model/proposal.py
@@ -15,7 +15,7 @@ from opengever.meeting.workflow import Workflow
 from opengever.ogds.base.utils import ogds_service
 from opengever.ogds.models import UNIT_ID_LENGTH
 from opengever.ogds.models import USER_ID_LENGTH
-from opengever.ogds.models.types import UnicodeCoercingText
+from opengever.base.types import UnicodeCoercingText
 from plone import api
 from sqlalchemy import Column
 from sqlalchemy import Date
